+++ src/Storage/Storage/Blob/Cmdlet/NewAzureStorageContainer.cs
@@ -83,6 +83,9 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob.Cmdlet
         }
         private bool? preventEncryptionScopeOverride;
 
+        // Overwrite the useless parameter
+        public override string TagCondition { get; set; }
+
         /// <summary>
         /// Initializes a new instance of the NewAzureStorageContainerCommand class.
         /// </summary>