+++ src/Storage/Storage.Management/ChangeLog.md
@@ -23,6 +23,8 @@
     -  `New-AzStorageContainerSASToken`
 * support create account Sas token with new permission x,t,f
     -  `New-AzStorageAccountSASToken`
+* Support get single file share usage
+    - `Get-AzRmStorageShare`
 * support get/set blob tags on a specific blob
     -  `Get-AzStorageBlobTag`
     -  `Set-AzStorageBlobTag`
@@ -43,7 +45,7 @@
     -  `Start-AzStorageBlobCopy`
     -  `Stop-AzStorageBlobCopy`
 
-
+        
 ## Version 2.3.0
 * Fixed the issue that UserAgent is not added for some data plane cmdlets.
 * Supported create/update Storage account with MinimumTlsVersion and AllowBlobPublicAccess